refactor: 자잘한 코드 수정
