+++ src/app/matching/mate-new-post/page.tsx
@@ -9,7 +9,7 @@ import { AxiosResponse } from 'axios';
 import { useRouter } from 'next/navigation';
 import { createMatePost } from '@/services/matching/postNewMatePost';
 import KakaoMapModal from '../components/KakaoMapModal';
-import { NewMateData } from '../../../types/matching/mateNewPostType';
+import { NewMateData } from '../../../types/matching/mateDataType';
 
 const MateNewPostPage = () => {
   const [isMapOpen, setIsMapOpen] = useState(false);
@@ -144,7 +144,7 @@ const MateNewPostPage = () => {
       {/* 날짜 선택 필드 */}
       <div className="mb-2.5">
         <div className="text-md font-bold">날짜</div>
-        <div className="rounded-md bg-gray-100 p-2 dark:bg-default-100">
+        <div className="rounded-medium bg-gray-100 p-2 dark:bg-default-100">
           <DatePicker
             dateFormat="YYYY년 MM월 dd일"
             selected={startDate}
